requirements.txt
psycopg2-binary
pydantic-settings
requests
python-dotenv
rapidfuzz
